refactor: split ably connection module
